README Update: Basics on how to use the tool.
I read through the README, and whoever reads it should know what the application contains and a basic description of its functionalities. Since I am new to your tool and I tried to use it (using the sampleS) during the writing of the file, I wonder if it's possible to add information to public genomic databases so users can explore the tool more effectively.